Move water schema code to its own service
